requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.7.0

# Transcription
openai-whisper>=20230314
torch>=2.0.0
torchaudio>=2.0.0

# Video processing
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
Pillow>=10.0.0

# AI/ML models
ollama>=0.1.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.66.0

# Optional: for CLIP-based filtering
transformers>=4.30.0
sentencepiece>=0.1.99